[Fixed rune counting issue when crafting using essense pouches] Fixed an issue that occurs when runecrafting using an essense pouch. It is possible empty a pouch while the runecrafting animation is playing, and then crafting the new rune essences before a different animation is played. This in turn fails to trigger the "onAnimationChanged()" event causing the plugin to miss the second set of runes.
To counter this the inventorySnapshot is only cleared when a new animation is detected. In addition, the processChange() code was altered to overwrite the inventorySnapshot with the currentInventory each time it is called. This means that ANY change to the inventory while the animation is playing will trigger the code and generate a new inventory diff, detecting any and all runes crafted. After the animation is changed, the code will work the same way as before.

[Added a "Reset All" option]
Added a "Reset All" context menu option to the topPanelItem. This allows the user to clear the plugins log of all previously crafted runes, and start fresh. (Code based on LootTrackerPanel.java from the Runelite repository).